+++ packages/@vssue/api-bitbucket-v2/src/index.ts
@@ -30,9 +30,7 @@ export default class BitbucketV2 implements VssueAPI.Instance {
   owner: string
   repo: string
   clientId: string
-  clientSecret: string
   state: string
-  proxy: string | ((url: string) => string)
   $http: AxiosInstance
 
   constructor ({
@@ -40,18 +38,14 @@ export default class BitbucketV2 implements VssueAPI.Instance {
     owner,
     repo,
     clientId,
-    clientSecret,
     state,
-    proxy,
   }: VssueAPI.Options) {
     this.baseURL = baseURL
     this.owner = owner
     this.repo = repo
 
     this.clientId = clientId
-    this.clientSecret = clientSecret
     this.state = state
-    this.proxy = proxy
 
     this.$http = axios.create({
       baseURL: 'https://api.bitbucket.org/2.0',
@@ -85,7 +79,8 @@ export default class BitbucketV2 implements VssueAPI.Instance {
     window.location.href = buildURL(concatURL(this.baseURL, 'site/oauth2/authorize'), {
       client_id: this.clientId,
       redirect_uri: window.location.href,
-      response_type: 'code',
+      response_type: 'token',
+      state: this.state,
     })
   }
 
@@ -97,53 +92,24 @@ export default class BitbucketV2 implements VssueAPI.Instance {
    * @see https://developer.atlassian.com/bitbucket/api/2/reference/meta/authentication#oauth-2
    *
    * @remarks
-   * If the `code` exists in the query, remove them from query, and try to get the access token.
+   * If the `access_token` and `state` exist in the query, and the `state` matches, remove them from query, and return the access token.
    */
   async handleAuth (): Promise<VssueAPI.AccessToken> {
-    const query = parseQuery(window.location.search)
-    if (query.code) {
-      const code = query.code
-      delete query.code
-      const replaceURL = buildURL(getCleanURL(window.location.href), query) + window.location.hash
-      window.history.replaceState(null, '', replaceURL)
-      const accessToken = await this.getAccessToken({ code })
-      return accessToken
+    const hash = parseQuery(window.location.hash.slice(1))
+    if (!hash.access_token || hash.state !== this.state) {
+      return null
     }
-    return null
-  }
-
-  /**
-   * Get user access token via `code`
-   *
-   * @param options.code - The code from the query
-   *
-   * @return User access token
-   *
-   * @see https://developer.atlassian.com/bitbucket/api/2/reference/meta/authentication#oauth-2
-   */
-  async getAccessToken ({
-    code,
-  }: {
-    code: string
-  }): Promise<string> {
-    const originalURL = concatURL(this.baseURL, 'site/oauth2/access_token')
-    const proxyURL = typeof this.proxy === 'function'
-      ? this.proxy(originalURL)
-      : this.proxy
-    const { data } = await this.$http.post(proxyURL, buildQuery({
-      grant_type: 'authorization_code',
-      redirect_uri: window.location.href,
-      code,
-    }), {
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-      },
-      auth: {
-        username: this.clientId,
-        password: this.clientSecret,
-      },
-    })
-    return data.access_token
+    const accessToken = hash.access_token
+    delete hash.access_token
+    delete hash.token_type
+    delete hash.expires_in
+    delete hash.state
+    delete hash.scopes
+    const hashString = buildQuery(hash)
+    const newHash = hashString ? `#${hashString}` : ''
+    const replaceURL = `${getCleanURL(window.location.href)}${window.location.search}${newHash}`
+    window.history.replaceState(null, '', replaceURL)
+    return accessToken
   }
 
   /**
@@ -257,7 +223,7 @@ export default class BitbucketV2 implements VssueAPI.Instance {
   }
 
   /**
-   * Get comments of this page according to the issue id or the issue title
+   * Get comments of this page according to the issue id
    *
    * @param options.accessToken - User access token
    * @param options.issueId - The id of issue
@@ -265,7 +231,7 @@ export default class BitbucketV2 implements VssueAPI.Instance {
    *
    * @return The comments
    *
-   * @see https://developer.atlassian.com/bitbucket/api/2/reference/resource/repositories/%7Busername%7D/%7Brepo_slug%7D/issues/%7Bissue_id%7D/comments#post
+   * @see https://developer.atlassian.com/bitbucket/api/2/reference/resource/repositories/%7Busername%7D/%7Brepo_slug%7D/issues/%7Bissue_id%7D/comments#get
    * @see https://developer.atlassian.com/bitbucket/api/2/reference/meta/pagination
    */
   async getComments ({
@@ -301,7 +267,7 @@ export default class BitbucketV2 implements VssueAPI.Instance {
       count: data.size,
       page: data.page,
       perPage: data.pagelen,
-      data: data.values.map(normalizeComment),
+      data: data.values.filter(item => item.content.raw !== null).map(normalizeComment),
     }
   }
 
@@ -397,14 +363,14 @@ export default class BitbucketV2 implements VssueAPI.Instance {
   /**
    * Bitbucket does not support reactions now
    */
-  async getCommentReactions (): Promise<VssueAPI.Reactions> {
+  async getCommentReactions (options): Promise<VssueAPI.Reactions> {
     throw new Error('501 Not Implemented')
   }
 
   /**
    * Bitbucket does not support reactions now
    */
-  async postCommentReaction (): Promise<boolean> {
+  async postCommentReaction (options): Promise<boolean> {
     throw new Error('501 Not Implemented')
   }
 }