+++ packages/vssue/src/i18n/index.ts
@@ -1,7 +1,9 @@
 import Vue from 'vue'
 import VueI18n from 'vue-i18n'
-import enUS from './lang/en-US'
-import zhCN from './lang/zh-CN'
+import enUS from './langs/en-US'
+import zhCN from './langs/zh-CN'
+import ptBR from './langs/pt-BR'
+import jaJP from './langs/ja-JP'
 
 if (!Vue.prototype.hasOwnProperty('$i18n')) {
   Vue.use(VueI18n)
@@ -15,6 +17,10 @@ const i18n: VueI18n = new VueI18n({
     'en-US': enUS,
     'zh': zhCN,
     'zh-CN': zhCN,
+    'pt': ptBR,
+    'pt-BR': ptBR,
+    'ja': jaJP,
+    'ja-JP': jaJP,
   },
 })